backend/requirements.txt
asgiref==3.10.0
bleach==6.1.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==24.3
django-js-asset==3.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
hypothesis==6.92.1
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
redis==5.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
gunicorn==21.2.0